backend/requirements.txt
beautifulsoup4==4.13.3
joblib==1.4.2
pandas==2.2.3
scikit_learn==1.6.1
ipykernel==6.29.5
seaborn==0.13.2
matplotlib==3.10.1
lxml==5.3.1
flask==3.1.0
pymongo==4.11.2
python-dotenv==1.0.1
gunicorn==23.0.0
xgboost==2.1.3
flask-cors==5.0.1
pytest==8.3.5
pytest-mock==3.14.0
torch==2.6.0